// xml-front/src/modules/patent/models/filter-dto.ts
export interface FilterDto {
  type: string | null
  operator: string | null
  value: string | null
}

/*

  numberOperator: string|null,
  number: string|null,
  titleOperator: string|null,
  title: string|null
  datumOperator: string|null
  datum: Date|null
  submitterOperator: string|null
  submitter: string|null
  inventorOperator: string|null
  inventor: string|null
  commissionerOperator: string|null
  commissioner: string|null
 */